libexfat: avoid overflow in file size calculation
ccount and clus_size are both 32-bit values, so their multiplication
can overflow before being stored in the 64-bit stream entry fields.
This can record an incorrect size when recovering a large contiguous
run of orphan clusters.

Promote ccount to uint64_t before the multiplication so the file size
is calculated using 64-bit arithmetic.

+++ lib/exfat_dir.c
@@ -763,8 +763,8 @@ int exfat_update_file_dentry_set(struct exfat *exfat,
 		}
 	}
 
-	dset[1].dentry.stream.valid_size = cpu_to_le64(ccount * exfat->clus_size);
-	dset[1].dentry.stream.size = cpu_to_le64(ccount * exfat->clus_size);
+	dset[1].dentry.stream.valid_size = cpu_to_le64((uint64_t)ccount * exfat->clus_size);
+	dset[1].dentry.stream.size = cpu_to_le64((uint64_t)ccount * exfat->clus_size);
 	if (start_clu)
 		dset[1].dentry.stream.start_clu = cpu_to_le32(start_clu);
